[IMP] convert to native javascript module
